test: storybook 디렉터리 수정 및 preview 변경
