+++ .storybook/preview.ts
@@ -1,13 +1,28 @@
 import type { Preview } from '@storybook/react';
 
+const customViewports = {
+  Default: {
+    name: 'Default',
+    styles: {
+      width: '415px',
+      height: '700px',
+    },
+  },
+};
+
 const preview: Preview = {
   parameters: {
     actions: { argTypesRegex: '^on[A-Z].*' },
+    layout: 'centered',
     controls: {
       matchers: {
         color: /(background|color)$/i,
         date: /Date$/,
       },
+    },
+    viewport: {
+      viewports: { ...customViewports },
+      defaultViewport: 'Default',
     },
   },
 };
