+++ src/components/Input/OwnerNameInput/OwnerNameInput.styles.ts
@@ -3,7 +3,7 @@ import { Container } from 'components/style/InputContainer';
 
 export const Styled = {
   OwnerNameInputWrapper: styled(Container)`
-    width: 100%;
+    width: 375px;
     margin-bottom: 16px;
     input {
       padding-left: 10px;
